chore(web): move provider components into settings subdirectory
