+++ apps/web/src/components/onboarding/onboarding-dialog.tsx
@@ -6,14 +6,14 @@ import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
 import { PROVIDER_META } from '@stitch/shared/providers/catalog';
 import { PROVIDER_IDS, type ProviderId } from '@stitch/shared/providers/types';
 
-import { FieldGroup, NoFieldsNote } from '@/components/provider-config/field-group';
+import { FieldGroup, NoFieldsNote } from '@/components/settings/providers/field-group';
 import {
   buildProviderConfigBody,
   hydrateProviderConfigState,
   resolveDefaultAuthMethod,
   type FieldValues,
-} from '@/components/provider-config/utils';
-import { ProviderLogo } from '@/components/settings/provider-logo';
+} from '@/components/settings/providers/utils';
+import { ProviderLogo } from '@/components/settings/providers/provider-logo';
 import { Button } from '@/components/ui/button';
 import { Dialog, DialogContent, DialogHeader, DialogTitle } from '@/components/ui/dialog';
 import { Input } from '@/components/ui/input';
